# Optimize performance by filtering dependencies early