Inspired by @Toflar’s #8275 this pull request tries to skip loading for packages that conflict with the root requirements. But after every [loading loop iteration](https://github.com/composer/composer/blob/4dabc17ec1e55d42d9611271d1728dcf9f79d56a/src/Composer/DependencyResolver/PoolBuilder.php#L65) in the `PoolBuilder` it extends the root requirements with new information.

This can be done because if all versions of a root dependency require the same package, this package can become a root dependency itself with a union of all version constraints. For example:

1. The root *composer.json* requires package `A` in version `‌^2.0`
2. The matching versions `2.0` and `2.1` get loaded, all lower versions get skipped.
3. `A 2.0` itself requires package `B` in `‌^3.0` and `A 2.1` requires `B` in `^3.1` therefore we can now add a root requirement for `B ^3.0‌`
4. Now when package `B` gets loaded all versions not matching `^3.0‌` get skipped directly.

I tested the version of my pull request and the results look very promising I think:

|                                      | 1.9.0                | 2.0-dev<br> 2019-08-02   | this pull request  |
|--------------------------------------|---------------------:|---------------------:|-------------------:|
| `"symfony/framework-bundle": "^3.4"` |  7.14 s <br> 265.14 MB |  3.82 s <br>  123.84 MB | 2.78 s <br>  67.03 MB |
| `"contao/managed-edition": "^4.8"`   | 33.10 s <br> 1618.04 MB | 87.53 s <br> 2802.62 MB | 8.02 s <br> 126.56 MB |
| `"contao/core-bundle": "^4.8"`       | 33.25 s <br> 1142.62 MB | 68.31 s <br> 2569.67 MB | 7.16 s <br> 77.73 MB |
| `"contao/core-bundle": "*"`          | 39.55 s <br> 1615.29 MB | 83.93 s <br> 2804.13 MB | 78.83 s <br> 2002.55 MB |

As you can see in the last example: if the requirement is very unrestricted the benefits aren’t that great, but I think in many real world use cases this change could make a huge impact.

The concept is pretty much the same as #8275 but without the need for manually adding root requires or conflicts.

The state of this pull request is just a first draft, if you like the concept I would continue working on it:

- [ ] Support `replaces` and `provide`.
- [ ] Support aliases.
- [ ] Add optimization to reduce the large number of constraints in the `MultiConstraint` objects.
- [ ] Update `loadPackages` method signature.
- [ ] Fix tests.