@Toflar 2019-08-28
I wonder why v2.0 is so much slower than 1.9 without your PR in some cases 🤔