@Seldaek 2019-08-28
Definitely looks like a good way to go to me, but tbh I'd wait before investing more time in this until #7936 gets completed as it's already huge and risks conflicting with a lot of things.