@Seldaek 2019-08-28
As for why 2.0 uses so much more memory and time than 1.x, that is indeed odd and should be investigated eventually but we also aren't there yet I think..