@ausi 2019-08-28
Thanks for the feedback. I’ll leave this PR as it is then for now.

I subscribed to #7936 so that I can continue working once it got merged. Otherwise feel free to ping me once you think 2.0 is ready for such pull requests ☺️