@ausi 2019-12-01
I will have to take another look at this, rebase the PR and compare the performance.