@Toflar 2019-12-02
I already did :) It's still very relevant but the current 2.0 branch still needs some updates. Optimizations should go to the new `PoolBuilder`. Let's work on this together, ping me on Slack :)