@stof 2020-11-26
should this be closed now that https://github.com/composer/composer/pull/8850 has been merged ?