@Seldaek 2020-11-26
Yup I don't think this makes a ton of sense anymore.